simple_trader/functions/trader/requirements.txt
cachetools==5.3.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
charset-normalizer==3.1.0 ; python_version >= "3.9" and python_version < "4"
click==8.1.3 ; python_version >= "3.9" and python_version < "4"
cloudevents==1.9.0 ; python_version >= "3.9" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4" and platform_system == "Windows"
datetime==5.1 ; python_version >= "3.9" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.9" and python_version < "4"
flask==2.2.3 ; python_version >= "3.9" and python_version < "4"
functions-framework==3.3.0 ; python_version >= "3.9" and python_version < "4"
google-api-core==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-core[grpc]==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.17.2 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-appengine-logging==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-audit-log==0.2.5 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-bigquery==3.9.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-logging==3.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-resumable-media==2.4.1 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.59.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos[grpc]==1.59.0 ; python_version >= "3.9" and python_version < "4.0"
grpc-google-iam-v1==0.12.6 ; python_version >= "3.9" and python_version < "4.0"
grpcio-status==1.53.0 ; python_version >= "3.9" and python_version < "4.0"
grpcio==1.53.0 ; python_version >= "3.9" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.9" and python_version < "4" and platform_system != "Windows"
idna==3.4 ; python_version >= "3.9" and python_version < "4"
importlib-metadata==6.1.0 ; python_version >= "3.9" and python_version < "3.10"
itsdangerous==2.1.2 ; python_version >= "3.9" and python_version < "4"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4"
markupsafe==2.1.2 ; python_version >= "3.9" and python_version < "4"
packaging==23.0 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.22.2 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.22.1 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.9" and python_version < "4"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
setuptools==67.6.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.9" and python_version < "4"
watchdog==3.0.0 ; python_version >= "3.9" and python_version < "4"
werkzeug==2.2.3 ; python_version >= "3.9" and python_version < "4"
zipp==3.15.0 ; python_version >= "3.9" and python_version < "3.10"
zope-interface==6.0 ; python_version >= "3.9" and python_version < "4.0"
